smo-cli/pyproject.toml
[project]
name = "smo-cli"
version = "0.1.0"
description = "A command-line interface for the Synergetic Meta-Orchestrator."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "smo-core",
    "click>=8.0",
    "rich>=12.0",
    "PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smo_cli"]

[project.scripts]
smo-cli = "smo_cli.cli:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "hatchling",
]

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-v -s"

[dependency-groups]
dev = [
    "abilian-devtools>=0.7.7",
    "coverage>=7.9.2",
]
